requirements.txt
pynput
pystray
Pillow
wmi
pywin32
